use single product block component for all country groups
